Fixing error with version number and carriage returns.
This bug was possibly dangerous in that it could cause DKMS to fail to build
all other modules (including video and network drivers) because a carriage
return symbol was inserted in the folder name and the dkms.conf file. If this
happened to you, you will come to a login screen and be unable to log in. To fix:
Drop to terminal with CTRL + ALT + F2. Log in as normal. Run sudo dkms autoinstall.
Once you are able to log in, remove the version folder from /var/lib/dkms and run
sudo dkms status to make sure it's gone. After that, go into /usr/src/
and fix the version number on the linuxcan folder and in the dkms.conf file in the folder.
Finally, to complete the install, run:
sudo dkms add linuxcan/<version>
sudo dkms build linuxcan/<version>
sudo dkms install linuxcan/<version>